refactor(observability): 抽 experience.ts skill frontmatter IO 到独立模块
把 experience.ts 里读 SKILL.md frontmatter 的三处投影(skill 类型 / 声明检查
/ expected_tools)抽到 src/observability/experience-frontmatter.ts:

- loadFrontmatterSkillType(skillName, cwd?) — 读 skillType / type / category
- loadSkillDeclarationCheck(skillName, cwd?) — 解析 hard rules / workflows
- loadExpectedToolsForSkill(skillName, cwd?) — 读 expected_tools

experience.ts 不再 import node:fs / hard-rules / skill-chain,主体回归
「事件 → 复盘字段」单职责,IO 改为通过 experience-frontmatter 显式调用。

阶段 3 PR2。experience.ts -75 行,新模块 105 行;Report 字段语义零变化。
